+++ tui/src/ui/components/playlist.rs
@@ -1,16 +1,25 @@
 use crate::ui::model::TermusicLayout;
-use crate::ui::tui_cmd::TuiCmd;
+use crate::ui::tui_cmd::{PlaylistCmd, TuiCmd};
 use crate::ui::Model;
-use anyhow::{anyhow, bail, Result};
+use anyhow::{anyhow, bail, Context as _, Result};
 use rand::seq::IndexedRandom;
 use std::borrow::Cow;
 use std::ffi::OsString;
 use std::path::Path;
+use termusiclib::config::v2::server::LoopMode;
 use termusiclib::config::SharedTuiSettings;
 use termusiclib::library_db::const_unknown::{UNKNOWN_ALBUM, UNKNOWN_ARTIST};
 use termusiclib::library_db::SearchCriteria;
 use termusiclib::library_db::TrackDB;
-use termusiclib::track::Track;
+use termusiclib::player::playlist_helpers::{
+    PlaylistAddTrack, PlaylistPlaySpecific, PlaylistRemoveTrackIndexed, PlaylistSwapTrack,
+    PlaylistTrackSource,
+};
+use termusiclib::player::{
+    PlaylistAddTrackInfo, PlaylistLoopModeInfo, PlaylistRemoveTrackInfo, PlaylistShuffledInfo,
+    PlaylistSwapInfo,
+};
+use termusiclib::track::{MediaType, Track};
 use termusiclib::types::{GSMsg, Id, Msg, PLMsg};
 use termusiclib::utils::{filetype_supported, get_parent_folder, is_playlist, playlist_get_vec};
 
@@ -262,12 +271,22 @@ impl Model {
     /// Add a playlist (like m3u) to the playlist.
     fn playlist_add_playlist(&mut self, current_node: &str) -> Result<()> {
         let vec = playlist_get_vec(current_node)?;
-        // "add_playlist" will still add all possible things, but return errors for which did not work
-        // so sync still needs to happen
-        let res = self.playlist.add_playlist(&vec);
-        self.player_sync_playlist()?;
-        self.playlist_sync();
-        res?;
+
+        let sources = vec
+            .into_iter()
+            .map(|v| {
+                if v.starts_with("http") {
+                    PlaylistTrackSource::Url(v)
+                } else {
+                    PlaylistTrackSource::Path(v)
+                }
+            })
+            .collect();
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::AddTrack(
+            PlaylistAddTrack::new_vec(u64::try_from(self.playlist.len()).unwrap(), sources),
+        )));
+
         Ok(())
     }
 
@@ -285,9 +304,12 @@ impl Model {
             .episodes
             .get(episode_index)
             .ok_or_else(|| anyhow!("get episode selected failed."))?;
-        self.playlist.add_episode(episode_selected);
-        self.player_sync_playlist()?;
-        self.playlist_sync();
+
+        let source = PlaylistTrackSource::PodcastUrl(episode_selected.url.clone());
+        self.command(TuiCmd::Playlist(PlaylistCmd::AddTrack(
+            PlaylistAddTrack::new_single(u64::try_from(self.playlist.len()).unwrap(), source),
+        )));
+
         Ok(())
     }
 
@@ -301,12 +323,16 @@ impl Model {
         }
         if p.is_dir() {
             let new_items_vec = Self::library_dir_children(p);
-            // "add_playlist" will still add all possible things, but return errors for which did not work
-            // so sync still needs to happen
-            let res = self.playlist.add_playlist(&new_items_vec);
-            self.player_sync_playlist()?;
-            self.playlist_sync();
-            res?;
+
+            let sources = new_items_vec
+                .into_iter()
+                .map(PlaylistTrackSource::Path)
+                .collect();
+
+            self.command(TuiCmd::Playlist(PlaylistCmd::AddTrack(
+                PlaylistAddTrack::new_vec(u64::try_from(self.playlist.len()).unwrap(), sources),
+            )));
+
             return Ok(());
         }
         self.playlist_add_item(current_node)?;
@@ -320,24 +346,29 @@ impl Model {
             self.playlist_add_playlist(current_node)?;
             return Ok(());
         }
-        self.playlist.add_track(&current_node)?;
-        self.player_sync_playlist()?;
+        let source = if current_node.starts_with("http") {
+            PlaylistTrackSource::Url(current_node.to_string())
+        } else {
+            PlaylistTrackSource::Path(current_node.to_string())
+        };
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::AddTrack(
+            PlaylistAddTrack::new_single(u64::try_from(self.playlist.len()).unwrap(), source),
+        )));
+
         Ok(())
     }
 
     /// Add [`TrackDB`] to the playlist
     pub fn playlist_add_all_from_db(&mut self, vec: &[TrackDB]) {
-        let vec2: Vec<&str> = vec.iter().map(|f| f.file.as_str()).collect();
-        // "add_playlist" will still add all possible things, but return errors for which did not work
-        // so sync still needs to happen
-        let res = self.playlist.add_playlist(&vec2);
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
-        }
-        self.playlist_sync();
-        if let Err(e) = res {
-            self.mount_error_popup(anyhow!(e).context("add all to playlist from database"));
-        }
+        let sources = vec
+            .iter()
+            .map(|f| PlaylistTrackSource::Path(f.file.clone()))
+            .collect();
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::AddTrack(
+            PlaylistAddTrack::new_vec(u64::try_from(self.playlist.len()).unwrap(), sources),
+        )));
     }
 
     /// Add random album(s) from the database to the playlist
@@ -364,6 +395,150 @@ impl Model {
             .get();
         let vec = self.playlist_get_random_tracks(playlist_select_random_track_quantity);
         self.playlist_add_all_from_db(&vec);
+    }
+
+    /// Handle when a playlist has added a track
+    pub fn handle_playlist_add(&mut self, items: PlaylistAddTrackInfo) -> Result<()> {
+        // piggyback off-of the server side implementation for now by re-parsing everything.
+        self.playlist.add_tracks(
+            PlaylistAddTrack {
+                at_index: items.at_index,
+                tracks: vec![items.trackid],
+            },
+            &self.podcast.db_podcast,
+        )?;
+
+        self.playlist_sync();
+
+        Ok(())
+    }
+
+    /// Handle when a playlist has removed a track
+    pub fn handle_playlist_remove(&mut self, items: &PlaylistRemoveTrackInfo) -> Result<()> {
+        let at_index = usize::try_from(items.at_index).unwrap();
+        // verify that it is the track to be removed via id matching
+        let Some(track_at_idx) = self.playlist.tracks().get(at_index) else {
+            // this should not happen as it is verified before the loop, but just in case
+            bail!("Failed to get track at index \"{at_index}\"");
+        };
+        // this unwrap could be handled better, but this should never actually happen
+        let id = track_at_idx.file().unwrap();
+
+        let input_track = &items.trackid;
+
+        // Note: clippy suggested this instead of a match block
+        let ((PlaylistTrackSource::Path(file_url), MediaType::Music)
+        | (PlaylistTrackSource::PodcastUrl(file_url), MediaType::Podcast)
+        | (PlaylistTrackSource::Url(file_url), MediaType::LiveRadio)) =
+            (&input_track, track_at_idx.media_type)
+        else {
+            bail!(
+                "Type mismatch, expected \"{:#?}\" at \"{at_index}\" found \"{:#?}\"",
+                input_track,
+                track_at_idx
+            );
+        };
+
+        if file_url != id {
+            bail!("URI mismatch, expected \"{file_url}\" at \"{at_index}\" in request, found \"{id}\" in playlist");
+        }
+
+        // piggyback off-of the server side implementation.
+        self.playlist.remove(at_index);
+
+        self.playlist_sync();
+
+        Ok(())
+    }
+
+    /// Handle when a playlist was cleared
+    pub fn handle_playlist_clear(&mut self) {
+        self.playlist.clear();
+
+        self.playlist_sync();
+    }
+
+    /// Handle when the playlist loop-mode was changed
+    pub fn handle_playlist_loopmode(&mut self, loop_mode: &PlaylistLoopModeInfo) -> Result<()> {
+        let as_u8 = u8::try_from(loop_mode.mode).context("Failed to convert u32 to u8")?;
+        let loop_mode =
+            LoopMode::tryfrom_discriminant(as_u8).context("Failed to get LoopMode from u8")?;
+        self.playlist.set_loop_mode(loop_mode);
+        self.config_server.write().settings.player.loop_mode = loop_mode;
+        self.playlist_update_title();
+        // Force a redraw as stream updates are not part of the "tick" event and so cant send "Msg"
+        // but need a redraw because ofthe title change
+        self.force_redraw();
+
+        Ok(())
+    }
+
+    /// Handle when the playlist had swapped some tracks
+    pub fn handle_playlist_swap_tracks(&mut self, swapped_tracks: &PlaylistSwapInfo) -> Result<()> {
+        let index_a = usize::try_from(swapped_tracks.index_a)
+            .context("Failed to convert index_a to usize")?;
+        let index_b = usize::try_from(swapped_tracks.index_b)
+            .context("Failed to convert index_b to usize")?;
+
+        self.playlist.swap(index_a, index_b)?;
+
+        self.playlist_sync();
+
+        Ok(())
+    }
+
+    /// Handle when the playlist has been shuffled and so has new order of tracks
+    pub fn handle_playlist_shuffled(&mut self, shuffled: PlaylistShuffledInfo) -> Result<()> {
+        let playlist_comp_selected_index = self.playlist_get_selected_index();
+        // this might be fragile if there are multiple of the same track in the playlist as there is no unique identifier currently
+        let playlist_track_at_old_index = playlist_comp_selected_index
+            .and_then(|idx| self.playlist.tracks().get(idx))
+            .and_then(|track| track.file())
+            .map(ToOwned::to_owned);
+
+        self.playlist
+            .load_from_grpc(shuffled.tracks, &self.podcast.db_podcast)?;
+        self.playlist_sync();
+
+        let found_new_index = self
+            .playlist
+            .tracks()
+            .iter()
+            .enumerate()
+            .find(|(_, track)| track.file() == playlist_track_at_old_index.as_deref());
+        if let Some((new_index, _)) = found_new_index {
+            self.playlist_locate(new_index);
+        }
+
+        Ok(())
+    }
+
+    /// Handle setting the current track index in the TUI playlist and selecting the proper list item
+    pub fn handle_current_track_index(&mut self, current_track_index: usize, force_relocate: bool) {
+        let tui_old_current_index = self.playlist.get_current_track_index();
+        info!(
+            "index from player is: {current_track_index:?}, index in tui is: {tui_old_current_index:?}"
+        );
+        self.playlist.clear_current_track();
+        self.playlist.set_current_track_index(current_track_index);
+
+        let playlist_comp_selected_index = self.playlist_get_selected_index();
+
+        // only re-select the current-track if the old selection was the old-current-track
+        if force_relocate || playlist_comp_selected_index.is_none_or(|v| v == tui_old_current_index)
+        {
+            self.playlist_locate(current_track_index);
+        }
+
+        self.current_song = self.playlist.current_track().cloned();
+        self.update_layout_for_current_track();
+        self.player_update_current_track_after();
+
+        self.lyric_update_for_podcast_by_current_track();
+
+        if let Err(e) = self.podcast_mark_current_track_played() {
+            self.mount_error_popup(e.context("Marking podcast track as played"));
+        }
     }
 
     fn playlist_sync_podcasts(&mut self) {
@@ -476,39 +651,91 @@ impl Model {
         self.playlist_update_title();
     }
 
+    /// Delete a track at `index` from the playlist
     pub fn playlist_delete_item(&mut self, index: usize) {
+        if self.playlist.is_empty() || index >= self.playlist.len() {
+            return;
+        }
+
+        let Some(item) = self.playlist.tracks().get(index) else {
+            return;
+        };
+
+        let Some(file) = item.file() else {
+            return;
+        };
+
+        let id = match item.media_type {
+            termusiclib::track::MediaType::Music => PlaylistTrackSource::Path(file.to_string()),
+            termusiclib::track::MediaType::Podcast => {
+                PlaylistTrackSource::PodcastUrl(file.to_string())
+            }
+            termusiclib::track::MediaType::LiveRadio => PlaylistTrackSource::Url(file.to_string()),
+        };
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::RemoveTrack(
+            PlaylistRemoveTrackIndexed::new_single(u64::try_from(index).unwrap(), id),
+        )));
+    }
+
+    /// Clear a entire playlist
+    pub fn playlist_clear(&mut self) {
         if self.playlist.is_empty() {
             return;
         }
-        self.playlist.remove(index);
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
-        }
-        self.playlist_sync();
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::Clear));
     }
 
-    pub fn playlist_clear(&mut self) {
-        self.playlist.clear();
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
-        }
-        self.playlist_sync();
-    }
-
+    /// Shuffle the whole playlist
     pub fn playlist_shuffle(&mut self) {
-        self.playlist.shuffle();
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
+        self.command(TuiCmd::Playlist(PlaylistCmd::Shuffle));
+    }
+
+    /// Send command to swap 2 indexes. Does nothing if either index is out-of-bounds.
+    ///
+    /// # Panics
+    ///
+    /// if `usize` cannot be converted to `u64`
+    fn playlist_swap(&mut self, index_a: usize, index_b: usize) {
+        let len = self.playlist.tracks().len();
+        if index_a.max(index_b) >= len {
+            error!(
+                "Index out-of-bounds, not executing swap: {}",
+                index_a.max(index_b)
+            );
+            return;
         }
-        self.playlist_sync();
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::SwapTrack(
+            PlaylistSwapTrack {
+                index_a: u64::try_from(index_a).unwrap(),
+                index_b: u64::try_from(index_b).unwrap(),
+            },
+        )));
+    }
+
+    /// Swap the given index upwards, does nothing if out-of-bounds or would result in itself.
+    pub fn playlist_swap_up(&mut self, index: usize) {
+        if index == 0 {
+            return;
+        }
+
+        // always guranteed to be above 0, no saturated necessary
+        self.playlist_swap(index, index - 1);
+    }
+
+    /// Swap the given index downwards, does nothing if out-of-bounds.
+    pub fn playlist_swap_down(&mut self, index: usize) {
+        if index >= self.playlist.len().saturating_sub(1) {
+            return;
+        }
+
+        self.playlist_swap(index, index.saturating_add(1));
     }
 
     pub fn playlist_update_library_delete(&mut self) {
-        self.playlist.remove_deleted_items();
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
-        }
-        self.playlist_sync();
+        self.command(TuiCmd::Playlist(PlaylistCmd::RemoveDeletedItems));
     }
 
     pub fn playlist_update_title(&mut self) {
@@ -536,12 +763,34 @@ impl Model {
             )
             .ok();
     }
+
+    /// Play the currently selected item in the playlist list
     pub fn playlist_play_selected(&mut self, index: usize) {
-        self.playlist.set_current_track_index(index);
-        if let Err(e) = self.player_sync_playlist() {
-            self.mount_error_popup(e.context("player sync playlist"));
-        }
-        self.command(TuiCmd::PlaySelected);
+        let Some(track) = self.playlist.tracks().get(index) else {
+            error!("Track {index} not in playlist!");
+            return;
+        };
+
+        // TODO: refactor Track::file to be always existing
+        let Some(file) = track.file() else {
+            error!("Track {index} did not have a file(id), skipping!");
+            return;
+        };
+        // TODO: this should likely be a function on "Track"
+        let id = match track.media_type {
+            termusiclib::track::MediaType::Music => PlaylistTrackSource::Path(file.to_string()),
+            termusiclib::track::MediaType::Podcast => {
+                PlaylistTrackSource::PodcastUrl(file.to_string())
+            }
+            termusiclib::track::MediaType::LiveRadio => PlaylistTrackSource::Url(file.to_string()),
+        };
+
+        self.command(TuiCmd::Playlist(PlaylistCmd::PlaySpecific(
+            PlaylistPlaySpecific {
+                track_index: u64::try_from(index).unwrap(),
+                id,
+            },
+        )));
     }
 
     pub fn playlist_update_search(&mut self, input: &str) {
@@ -549,6 +798,7 @@ impl Model {
         self.general_search_update_show(Model::build_table(filtered_music));
     }
 
+    /// Select the given index in the playlist list component
     pub fn playlist_locate(&mut self, index: usize) {
         assert!(self
             .app
@@ -558,6 +808,17 @@ impl Model {
                 AttrValue::Payload(PropPayload::One(PropValue::Usize(index))),
             )
             .is_ok());
+    }
+
+    /// Get the current selected index in the playlist list component
+    pub fn playlist_get_selected_index(&self) -> Option<usize> {
+        // the index on a "Table" can be set via "AttrValue::Payload(PropPayload::One(PropValue::Usize(val)))", but reading that is stale
+        // as that value is only read in the "Table", not removed or updated, but "state" is
+        let Ok(State::One(StateValue::Usize(val))) = self.app.state(&Id::Playlist) else {
+            return None;
+        };
+
+        Some(val)
     }
 
     pub fn playlist_get_random_tracks(&mut self, quantity: u32) -> Vec<TrackDB> {